lang & tooltips
