+++ src/main/java/co/neeve/nae2/common/blocks/BlockExposer.java
@@ -6,7 +6,8 @@ import net.minecraft.block.material.Material;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
 import java.util.List;
 
@@ -16,31 +17,10 @@ public class BlockExposer extends AEBaseTileBlock {
 	}
 
 	@Override
+	@SideOnly(Side.CLIENT)
 	public void addInformation(ItemStack is, World world, List<String> lines, ITooltipFlag advancedItemTooltips) {
 		super.addInformation(is, world, lines, advancedItemTooltips);
 
-		lines.add("Exposes the network contents as capabilities.");
-
-		var registered = NAE2.api().exposer().getRegisteredHandlers();
-		if (registered.isEmpty()) {
-			lines.add("");
-			lines.add("No handlers registered.");
-		} else {
-			lines.add("");
-			lines.add("Registered handlers:");
-
-			for (var handler : registered.object2ObjectEntrySet()) {
-				var name = handler.getKey().getName();
-
-				// If name is a class path, strip everything but the name.
-				if (name.contains(".")) {
-					name = name.substring(name.lastIndexOf('.') + 1);
-				}
-
-				lines.add(" - "
-					+ "§6" + name + "§r"
-					+ " (" + handler.getValue().getMod().getAnnotation(Mod.class).name() + ")");
-			}
-		}
+		NAE2.api().exposer().addTooltipInformation(is, world, lines, advancedItemTooltips);
 	}
 }
